Pagina web funeraria Backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
sqlalchemy==2.0.48
bcrypt==5.0.0
python-jose[cryptography]==3.5.0
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.2
python-multipart==0.0.22
email-validator==2.3.0
pytest==9.0.2
httpx==0.28.1
